Final changes
Fixed some bugs like : admin can delete himself , empty error massage in create transaction , chagned privacy policy , create default admin account and automated data migration when running application

+++ SmartWallet/SmartWallet/Controllers/AdminController.cs
@@ -50,16 +50,24 @@ namespace SmartWallet.Controllers
         }
 
         [HttpGet]
-        public async Task<IActionResult> Transactions()
+        public async Task<IActionResult> Transactions(TransactionStatus? status)
         {
             var user = await _userManager.GetUserAsync(User);
             if (user is null) return NotFound();
 
-            var transactions = _context.Transactions
+            var query = _context.Transactions
             .Include(t => t.Sender)
             .Include(t => t.Receiver)
+            .AsQueryable();
+
+            if (status.HasValue)
+            {
+                query = query.Where(t => t.Status == status.Value); 
+            }
+            var transactions = await query.OrderByDescending(t => t.Date).ToListAsync();
+
+            ViewBag.CurrentFilter = status;
             
-            .ToList();
             return View(transactions);
         }
         [HttpGet]
@@ -77,26 +85,38 @@ namespace SmartWallet.Controllers
         public async Task<IActionResult> TransactionStatusSetToComplete(int id)
         {
             var transaction = await _context.Transactions
+                .Include(t => t.Sender)
+                .Include(t => t.Receiver)
                 .FirstOrDefaultAsync(t => t.TransactionID == id);
             if (transaction is null) return NotFound();
             
             if(transaction.Status != TransactionStatus.Completed)
             {
                 transaction.Status = TransactionStatus.Completed;
+                transaction.Sender.Balance -= transaction.Amount;
+                transaction.Receiver.Balance += transaction.Amount;
+
+                await _context.SaveChangesAsync();
 
             }
             return RedirectToAction(nameof(Transactions));
         }
         [Authorize(Roles="Admin")]
         public async Task<IActionResult> TransactionStatusSetToCancel(int id)
-        {
-            var transaction = await _context.Transactions.FindAsync(id);
+        {  
+            var transaction = await _context.Transactions
+                .Include(t => t.Sender)
+                .Include(t => t.Receiver)
+                .FirstOrDefaultAsync(t => t.TransactionID == id);
 
             if (transaction is null) return NotFound();
 
             if(transaction.Status != TransactionStatus.Cancelled)
             {
                 transaction.Status = TransactionStatus.Cancelled;
+                transaction.Sender.Balance += transaction.Amount;
+                transaction.Receiver.Balance -= transaction.Amount;
+                await _context.SaveChangesAsync();
             }
             return RedirectToAction(nameof(Transactions));
         }
@@ -113,11 +133,22 @@ namespace SmartWallet.Controllers
         public async Task<IActionResult> DeleteUser(string id)
         {
             var user = await _userManager.FindByIdAsync(id);
+            var currUserID = _userManager.GetUserId(User);
             if (user is null) return NotFound();
+
+            if( id == currUserID)
+            {
+                TempData["Error"] =  "Nie można usunąć własnego konta.";
+                return RedirectToAction("Users");
+            }
+            var userTransactions = _context.Transactions
+                .Where(t => t.SenderId == id ||  t.ReceiverId == id);
+            _context.Transactions.RemoveRange(userTransactions);
 
             var result = await _userManager.DeleteAsync(user);
             if (result.Succeeded)
             {
+                TempData["Success"] = "Użytkownik został pomyślnie usunięty.";
                 return RedirectToAction("Users");
             }
             else
